updated to release dependencies and log changed to error

+++ source_code/server/com.nagoya.middleware/src/main/java/com/nagoya/middleware/service/ContractResourceService.java
@@ -6,19 +6,25 @@ package com.nagoya.middleware.service;
 
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Iterator;
 import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.Session;
 
+import com.nagoya.dao.base.BasicDAO;
+import com.nagoya.dao.base.impl.BasicDAOImpl;
 import com.nagoya.dao.contract.ContractDAO;
 import com.nagoya.dao.contract.impl.ContractDAOImpl;
 import com.nagoya.dao.util.StringUtil;
+import com.nagoya.middleware.main.ServerPropertiesProvider;
+import com.nagoya.middleware.main.ServerProperty;
 import com.nagoya.middleware.util.DefaultDateProvider;
 import com.nagoya.middleware.util.DefaultIDGenerator;
 import com.nagoya.middleware.util.DefaultReturnObject;
 import com.nagoya.model.dbo.contract.ContractDBO;
+import com.nagoya.model.dbo.contract.ContractFileDBO;
 import com.nagoya.model.dbo.contract.Status;
 import com.nagoya.model.dbo.user.OnlineUserDBO;
 import com.nagoya.model.dbo.user.RequestType;
@@ -30,8 +36,11 @@ import com.nagoya.model.exception.InvalidObjectException;
 import com.nagoya.model.exception.InvalidTokenException;
 import com.nagoya.model.exception.NonUniqueResultException;
 import com.nagoya.model.exception.NotAuthorizedException;
+import com.nagoya.model.exception.NotFoundException;
 import com.nagoya.model.exception.ResourceOutOfDateException;
 import com.nagoya.model.exception.TimeoutException;
+import com.nagoya.model.to.contract.ContractFileTO;
+import com.nagoya.model.to.contract.ContractFileTransformer;
 
 /**
  * @author Florin Bogdan Balint
@@ -52,7 +61,7 @@ public class ContractResourceService extends ResourceService {
     public DefaultReturnObject create(String authorization, String language, com.nagoya.model.to.contract.ContractTO contractTO)
         throws NotAuthorizedException, ConflictException, TimeoutException, InvalidTokenException, InvalidObjectException, ResourceOutOfDateException,
         BadRequestException, NonUniqueResultException {
-        LOGGER.debug("Adding genetic resource");
+        LOGGER.debug("Adding contract");
         OnlineUserDBO onlineUser = validateSession(authorization);
 
         if (contractTO == null) {
@@ -64,8 +73,8 @@ public class ContractResourceService extends ResourceService {
         com.nagoya.model.dbo.contract.ContractDBO contractDBO = contractFactory.createDBOContract(onlineUser.getPerson(), contractTO);
 
         Calendar cal = Calendar.getInstance();
-        // TODO: externalize to config
-        cal.add(Calendar.DAY_OF_YEAR, 30);
+        int integer = ServerPropertiesProvider.getInteger(ServerProperty.CONTRACT_EXPIRAETION_TIME, 72);
+        cal.add(Calendar.HOUR, integer);
 
         Date expirationDate = cal.getTime();
         String token = DefaultIDGenerator.generateRandomID();
@@ -105,6 +114,14 @@ public class ContractResourceService extends ResourceService {
         }
 
         ContractDBO contractToDelete = (ContractDBO) contractDAO.find(contractIdLong, ContractDBO.class);
+
+        // verify sender
+        long longValueUser = onlineUser.getPerson().getId().longValue();
+        long longValueSender = contractToDelete.getSender().getId().longValue();
+        if (longValueUser != longValueSender) {
+            throw new ForbiddenException("Cannot delete foreign contract.");
+        }
+
         Status status = contractToDelete.getStatus();
         if (status.equals(Status.CREATED)) {
             contractDAO.delete(contractToDelete, true);
@@ -137,6 +154,126 @@ public class ContractResourceService extends ResourceService {
         LOGGER.debug("Searching for contracts");
         List<ContractDBO> results = contractDAO.search(onlineUser.getPerson(), dateFromToFilter, dateUntilToFilter, statusToFilter, 0);
         LOGGER.debug("Results found: " + results.size());
+
+        DefaultReturnObject result = refreshSession(onlineUser, null, null);
+        return result;
+    }
+
+    public DefaultReturnObject create(String authorization, String language, String contractId, ContractFileTO contractFileTO)
+        throws NotAuthorizedException, ConflictException, TimeoutException, InvalidTokenException, BadRequestException, NonUniqueResultException,
+        InvalidObjectException, ResourceOutOfDateException, ForbiddenException {
+        LOGGER.debug("Adding file to contract resource");
+        OnlineUserDBO onlineUser = validateSession(authorization);
+
+        if (contractFileTO == null) {
+            throw new BadRequestException("Cannot execute create operation, when the object is null.");
+        }
+
+        if (StringUtil.isNullOrBlank(contractId)) {
+            throw new BadRequestException("Contract ID must be provided.");
+        }
+
+        long contractIdLong = 0;
+        try {
+            contractIdLong = Long.parseLong(contractId);
+        } catch (NumberFormatException e) {
+            throw new BadRequestException("Contract ID must be a number.");
+        }
+
+        ContractDBO contractDBO = (ContractDBO) contractDAO.find(contractIdLong, ContractDBO.class);
+
+        // verify sender
+        long longValueUser = onlineUser.getPerson().getId().longValue();
+        long longValueSender = contractDBO.getSender().getId().longValue();
+        if (longValueUser != longValueSender) {
+            throw new ForbiddenException("Cannot add file to foreign contract.");
+        }
+
+        ContractFileDBO contractFileDBO = ContractFileTransformer.getDBO(contractFileTO);
+        contractDBO.getFiles().add(contractFileDBO);
+
+        contractDAO.update(contractDBO, true);
+
+        DefaultReturnObject result = refreshSession(onlineUser, null, null);
+        return result;
+    }
+
+    public DefaultReturnObject retrieveFile(String authorization, String language, String contractId, String fileId)
+        throws BadRequestException, NotAuthorizedException, ConflictException, TimeoutException, InvalidTokenException, NonUniqueResultException,
+        InvalidObjectException, ResourceOutOfDateException, NotFoundException {
+        OnlineUserDBO onlineUser = validateSession(authorization);
+
+        long fileIdLong = 0;
+        try {
+            fileIdLong = Long.parseLong(fileId);
+        } catch (NumberFormatException e) {
+            throw new BadRequestException("Contract ID must be a number.");
+        }
+
+        // TODO: limit contract file access if necessary...
+
+        BasicDAO<ContractFileDBO> basicDAO = new BasicDAOImpl<ContractFileDBO>(session);
+        ContractFileDBO contractFileDBO = (ContractFileDBO) basicDAO.find(fileIdLong, ContractFileDBO.class);
+        if (contractFileDBO == null) {
+            throw new NotFoundException("No file found");
+        }
+
+        ContractFileTO to = ContractFileTransformer.getTO(contractFileDBO);
+
+        DefaultReturnObject result = refreshSession(onlineUser, to, null);
+        return result;
+    }
+
+    public DefaultReturnObject deleteFile(String authorization, String language, String contractId, String fileId)
+        throws NotAuthorizedException, ConflictException, TimeoutException, InvalidTokenException, BadRequestException, ForbiddenException,
+        InvalidObjectException, ResourceOutOfDateException, NonUniqueResultException, NotFoundException {
+        LOGGER.debug("Adding file to contract resource");
+        OnlineUserDBO onlineUser = validateSession(authorization);
+
+        if (StringUtil.isNullOrBlank(contractId) || StringUtil.isNullOrBlank(fileId)) {
+            throw new BadRequestException("Contract and file ID must be provided.");
+        }
+
+        long contractIdLong = 0;
+        try {
+            contractIdLong = Long.parseLong(contractId);
+        } catch (NumberFormatException e) {
+            throw new BadRequestException("Contract ID must be a number.");
+        }
+
+        long fileIdLong = 0;
+        try {
+            fileIdLong = Long.parseLong(fileId);
+        } catch (NumberFormatException e) {
+            throw new BadRequestException("File ID must be a number.");
+        }
+
+        ContractDBO contractDBO = (ContractDBO) contractDAO.find(contractIdLong, ContractDBO.class);
+        if (contractDBO == null) {
+            throw new NotFoundException("Contract not found.");
+        }
+
+        // verify sender
+        long longValueUser = onlineUser.getPerson().getId().longValue();
+        long longValueSender = contractDBO.getSender().getId().longValue();
+        if (longValueUser != longValueSender) {
+            throw new ForbiddenException("Cannot add file to foreign contract.");
+        }
+
+        if (!contractDBO.getStatus().equals(Status.CREATED)) {
+            throw new ForbiddenException("File cannot be deleted any more.");
+        }
+
+        Iterator<ContractFileDBO> iterator = contractDBO.getFiles().iterator();
+        while (iterator.hasNext()) {
+            ContractFileDBO file = iterator.next();
+            if (file.getId().longValue() == fileIdLong) {
+                iterator.remove();
+                break;
+            }
+        }
+
+        contractDAO.update(contractDBO, true);
 
         DefaultReturnObject result = refreshSession(onlineUser, null, null);
         return result;
